Add an "IApplicationBuilder" to the project and implement required functionality.
So there are several updates here. I decided to add a "Builder" class to parse
arguments as they came in. This was necessary so I could test some but not all
command line behaviors without a big mess (it's hard to test static methods).

To do this I needed to make a view factory. I decided to make this take both a
string name and an appendable. I decided to not make it take in a model because
then I could make it a method. As a setter method, we get the advantage of
being able to change the model around a bunch with the controller maybe by
making a new instance and still be able to use the same window with the user.
An example we talked about was having the controller create a "Layered"
instance where maybe it only exposes a layer of objects to the view.
I of course added tests for everything I was implementing here. I think I still
need to flesh out the main method to do command line parsing, which luckly we
don't need to test.