Create main method command line parsing code.
This main method parses command line arguments really simply. It simply looks
for pairs for values and calls the right builder methods. If an error occurs
then it creates a JOptionPane just like the spec talks about.

One downside of creating of our own argument parser is weird behavior. Like you
can specify many of the same argument, and the behavior will be to override.

I also added a file to the base directory. I created a test configuration for
myself that used this file and the program ran! See
https://www.jetbrains.com/help/idea/run-debug-configuration-application.html#
for an article about how to set the command line arguments for yourself. Look
under the section called program arguments.