Implement the IAnimatorModelBuilder and cry about the tech debt incurred.
Since our model doesn't have keyframes and loosely decouples motions from shapes
there is some weirdness in the builder.

The biggest consequences I can think of are that shapes without motions have
a default set of properties. And a shape cannot be defined with a mix of
motions and keyframes (something keyframe systems don't have to deal with).

We've been had!