Fix test suite by adding canvas, and removing annoying swing window creation.
By getting a view in the ViewFactory I unfortunately triggered the construction
of a JFrame. This puts an icon in the dock even if we don't make the window
visible. I don't know the behavior on Linux or Windows but on Mac this was
really annoying. So I just disabled the test. I don't expect it is a biggie
since we will be manually verifying that the swing view is there anyway.

Additionally, I adjusted a test that was chronically failing. That was because
I was asking it to create a document from a blank input file. A model needs
at least a canvas in order to work with our views. So that was fixed.