Fix comments in AnimationModelBuilder and update B.App.Builder to file write.
Some comments talked about using Ovals to track state. But I actually use a
private class to get around that hack. So I removed that comment since it was
wrong.

Next, and this is the major one ... I fixed the BasicApplicationBuilder to
actually write to a file. This is something that really stumped me. The
solution, though I didn't like it initially because it wasn't elegant, was
to actually track filewriters and filereaders if we created them and close them
after making the view visible.

I believe because I wasn't closing the file, my changes which were small, might
have gotten stuck in a buffer and never flushed.
Now we flush changes and files are definitely written to. I verified this
manually since I want to avoid too much io testing. Who knows what the handins
server is going to say when I try to create a temporary file.