Add a view controller with binding logic stuck in the editor window.
Look at the action listeners I use. I tried to build it out so if we wanted to
move more logic to the controller then we could, though we would need to set
some methods on the view to get some information like selected shape and stuff.

The private helpers to update fields could be made public.

The assignment spec does seem to suggest a lot of controller control.
But I personally understand the view better this way. It controls its own
buttons and ferries user requests to the model (and does complicated logic
like renaming a shape which is something a user might want to do).