Add scrubbing panel and its fields, constructor, and important methods.
I added a scrubbing panel to represent the scrubbing component. It uses a change listener to update the controller whenever something moves. It also has set scrubbing max method for when the max value of a model changes. Additionally,it has a set position method to use whilethe animation is playing.